refactor: extract LoginPrompt.js into own file and use in /spacdeck
