+++ pages/spacedeck/[[...roomId]].js
@@ -20,6 +20,7 @@ import ServiceLink from '../../components/UI/ServiceLink';
 import CreateNewSketch from './actions/CreateNewSketch';
 import AddExistingSketch from './actions/AddExistingSketch';
 import { path as spacedeckPath } from '../../lib/Spacedeck';
+import LoginPrompt from '../../components/UI/LoginPrompt';
 
 export default function Spacedeck() {
     const auth = useAuth();
@@ -109,7 +110,6 @@ export default function Spacedeck() {
                 logger.debug(error);
                 setIsSpacedeckServerDown(true);
             });
-
             // Update the Matrix structure based on spacedeck sketches
             syncSketches && await updateStructure(spacedeck.getStructure());
             setSyncingServerSketches(false);
@@ -188,7 +188,7 @@ export default function Spacedeck() {
                     ]}
                 />
                 { errorMessage && <ErrorMessage>{ errorMessage }</ErrorMessage> }
-                { !auth.connectionStatus.spacedeck && 'insert login prompt' }
+                { !auth.connectionStatus.spacedeck && <LoginPrompt /> }
                 { !serviceSpaceId || syncingServerSketches ?
                     <LoadingSpinner /> :
                     <>
